requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
alembic==1.13.2
psycopg[binary]==3.2.3
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
PyJWT==2.9.0
argon2-cffi==23.1.0
httpx==0.27.2
python-ulid==2.7.0
orjson==3.10.7